+++ app/src/main/java/org/simple/clinic/splash/SplashScreen.kt
@@ -3,12 +3,21 @@ package org.simple.clinic.splash
 import android.content.Context
 import android.util.AttributeSet
 import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.navigation.findNavController
 import com.airbnb.lottie.LottieDrawable
+import kotlinx.android.parcel.Parcelize
 import org.simple.clinic.R
 import org.simple.clinic.databinding.ScreenSplashBinding
+import org.simple.clinic.di.injector
+import org.simple.clinic.navigation.v2.Router
+import org.simple.clinic.navigation.v2.compat.wrap
+import org.simple.clinic.onboarding.OnboardingScreen.OnboardingScreenKey
+import org.simple.clinic.router.screen.FullScreenKey
+import javax.inject.Inject
 
 class SplashScreen(context: Context, attributeSet: AttributeSet) : ConstraintLayout(context, attributeSet) {
+
+  @Inject
+  lateinit var router: Router
 
   private var binding: ScreenSplashBinding? = null
 
@@ -25,6 +34,8 @@ class SplashScreen(context: Context, attributeSet: AttributeSet) : ConstraintLay
       return
     }
 
+    context.injector<Injector>().inject(this)
+
     with(splashLottieView) {
       setAnimation("splash_animation.json")
       repeatCount = LottieDrawable.INFINITE
@@ -32,12 +43,24 @@ class SplashScreen(context: Context, attributeSet: AttributeSet) : ConstraintLay
     }
 
     nextButton.setOnClickListener {
-      findNavController().navigate(R.id.action_splashScreen_to_onboardingScreen)
+      router.clearHistoryAndPush(OnboardingScreenKey.wrap())
     }
   }
 
   override fun onDetachedFromWindow() {
     binding = null
     super.onDetachedFromWindow()
+  }
+
+  interface Injector {
+    fun inject(target: SplashScreen)
+  }
+
+  @Parcelize
+  object SplashScreenKey : FullScreenKey {
+
+    override val analyticsName = "Splash Screen"
+
+    override fun layoutRes() = R.layout.screen_splash
   }
 }