+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Convert `screen_patient_summary` to use `ConstraintLayout`
 - Bump Play Core to v1.10.0
 - Add option to manually trigger GH Actions
+- Migrate `SetupActivity` to the new navigation framework
 - [In Progress: 20 Jan 2021] Material Theme-ing Migration
 - [In Progress: 08 Feb 2021] Migrate app to use ViewBinding